Fix warnings

+++ Os/Posix/FileSystem.cpp
@@ -45,7 +45,7 @@ PosixFileSystem::Status PosixFileSystem::_rename(const char* originPath, const c
 
 PosixFileSystem::Status PosixFileSystem::_getWorkingDirectory(char* path, FwSizeType bufferSize) {
     Status status = OP_OK;
-    if (::getcwd(path, bufferSize) == nullptr) {
+    if (::getcwd(path, static_cast<size_t>(bufferSize)) == nullptr) {
         status = errno_to_filesystem_status(errno);
     }
     return status;
